% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowstate_utils.R
\name{add.keywords.to.data}
\alias{add.keywords.to.data}
\title{Add \verb{[['keyword']]} values to \verb{[['data']]}}
\usage{
add.keywords.to.data(flowstate, keywords.to.add)
}
\arguments{
\item{flowstate}{A flowstate object as returned from \link{read.flowstate}.}

\item{keywords.to.add}{Character vector; keyword names in \verb{[['keywords']]} whose factored values will be added to \verb{[['data']]}.}
}
\value{
UPDATES BY REFERENCE:
\itemize{
\item \verb{[['data']]}; factored \code{keywords.to.add} are added as additional columns.
\item \verb{[['keywords']]}; \code{keywords.to.add} are converted to class \code{factor}.
}

Invisibly returns \code{flowstate}.
}
\description{
Used for efficient addition of factored keyword values to \verb{[['data']]}; updates by reference.
}
\examples{
fcs.file.paths <- system.file("extdata", package = "flowstate") |>
list.files(full.names = TRUE, pattern = "BLOCK.*.fcs")

#read all .fcs files as flowstates; concatenate into a single object
fs <- read.flowstate(
  fcs.file.paths,
  colnames.type = "S",
  concatenate = TRUE
)

#create a new set of keywords/values
fs$keywords[
,
j = c('block.id', 'block.aliquot') := data.table::tstrsplit(
sample.id, "_", keep = 4:5)
]
#new columns; character class
fs$keywords[, .(block.id, block.aliquot)]

#add the factored keyword values to fs[['data']]
add.keywords.to.data(fs, c('block.id', 'block.aliquot'))

fs$data[, .(block.id, block.aliquot)]

#add additional keywords to [['data']]
add.keywords.to.data(fs, c('$DATE', '$PROJ'))

fs$data[, .(`$DATE`, `$PROJ`)]
fs$keywords[, .(`$DATE`, `$PROJ`)]

}
