% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowstate-package.R
\docType{package}
\name{flowstate-package}
\alias{flowstate-package}
\title{flowstate: Flow Cytometry Standard (FCS) Processing, Analysis, and Visualization}
\description{
Flow cytometry standard (.fcs) files are S3 structured to facilitate efficient processing, analysis, and visualization of data -- acheive your flow state.
}
\author{
\strong{Maintainer}: Nathan Laniewski \email{nathan_laniewski@urmc.rochester.edu}

}
\keyword{internal}
