% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowstate.R
\name{flowstate}
\alias{flowstate}
\title{flowstate: read, process, and store .fcs data}
\usage{
flowstate(
  fcs.file.path,
  colnames.type = c("S_N", "S", "N"),
  cofactor = 5000,
  sample.id = "TUBENAME"
)
}
\arguments{
\item{fcs.file.path}{Character string (length 1); full path to .fcs file.}

\item{colnames.type}{Character string; one of:
\itemize{
\item \code{"S_N"} -- \verb{[['data']]} columns are named by combining $PS (stain) and $PN (name), separated by an underscore.
\item \code{"S"} -- \verb{[['data']]} columns are named by using only their respective $PS (stain) keyword value.
\item \code{"N"} -- \verb{[['data']]} columns are named by using only their respective $PN (name) keyword value.
}}

\item{cofactor}{Numeric; default \code{5000}. Any/all parameters with a \verb{$PnTYPE} of 'Raw_Fluorescence' or 'Unmixed_Fluorescence' will be transformed using \link{asinh} and the defined cofactor value (\code{asinh(x/cofactor)}).}

\item{sample.id}{Character string; the keyword label defined through \code{sample.id} (default \code{TUBENAME}) will be used to add respective keyword values from \verb{[['keywords']]} as an identifier to \verb{[['data']]}.}
}
\value{
An object of class flowstate.
}
\description{
flowstate: read, process, and store .fcs data
}
