% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_flowstate.R
\name{flowstate.transform.inverse}
\alias{flowstate.transform.inverse}
\title{Transform values to linear -- inverse}
\usage{
flowstate.transform.inverse(flowstate, j = NULL)
}
\arguments{
\item{flowstate}{A flowstate object as returned from \link{read.flowstate}.}

\item{j}{Character vector -- default \code{NULL}; any/all parameters having a stored \code{transformed} attribute -- from the result of \link{flowstate.transform} -- will be inverse transformed in \verb{[['data']]} using the stored function (default \code{\link[base:Hyperbolic]{base::sinh()}}).  If a defined character vector: specific columns in \verb{[['data']]} that are to be inverse transformed.}
}
\value{
UPDATES BY REFERENCE:
\itemize{
\item \code{flowstate[['data']]}; inverse transformed values -- linear.
\item \code{flowstate[['data']]}; the stored attribute -- \code{transformed} -- is removed from each \code{j}.
}

Invisibly returns \code{flowstate}.
}
\description{
Transform values to linear -- inverse
}
\examples{
fcs.file.paths <- system.file("extdata", package = "flowstate") |>
list.files(full.names = TRUE, pattern = "BLOCK.*.fcs")

#read .fcs files as a flowstate object; concatenate
fs <- read.flowstate(
  fcs.file.paths,
  colnames.type = "S",
  concatenate = TRUE
)

#plot and mean values of linear columns
plot(fs, CD3, CD8) + ggplot2::guides(fill = 'none')
res1.linear <- fs$data[, sapply(.SD, mean), .SDcols = c('CD3', 'CD8')]
print(res1.linear)

#transform
flowstate.transform(
  fs,
  j = c('CD3','CD4','CD8'),
  transform.func = "asinh",
  cofactor = 5000
)

#updated [['data']] attributes; applied transform function and cofactor
fs$data[, lapply(.SD, attr, which = 'transformed')]
fs$data[, attr(CD3, 'transformed')]

#plot and mean values of transformed columns from updated fs[['data']]
plot(fs, CD3, CD8) + ggplot2::guides(fill = 'none')
fs$data[, sapply(.SD, mean), .SDcols = c('CD3', 'CD4', 'CD8')]

#inverse transformation; defined `j` argument
flowstate.transform.inverse(fs, j = c('CD3', 'CD8'))

#updated [['parameters']] attribute
#returns only `PerCP-Fire 806-A (CD4)` as all other transformations have been inversed
'transformed' \%in\% names(attributes(fs$parameters))
attr(fs$parameters, 'transformed')

#plot and mean values of linear columns
plot(fs, CD3, CD8) + ggplot2::guides(fill = 'none')
res2.linear <- fs$data[, sapply(.SD, mean), .SDcols = c('CD3', 'CD8')]
print(res2.linear)

#linear --> transformed --> inverse --> linear
res1.linear == res2.linear

}
