% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_flowstate.R
\name{flowstate.transform}
\alias{flowstate.transform}
\title{Transform linear values}
\usage{
flowstate.transform(
  flowstate,
  j = NULL,
  transform.type = "asinh",
  cofactor = 5000
)
}
\arguments{
\item{flowstate}{A flowstate object as returned from \link{read.flowstate}.}

\item{j}{Character vector -- default \code{NULL}; any/all parameters having a keyword-value pair of \code{TYPE/Raw|Unmixed_Fluorescence} will be transformed in \verb{[['data']]}.  If a defined character vector: specific columns in \verb{[['data']]} that are to be transformed.}

\item{transform.type}{Character vector -- default \link{asinh}; quoted function that will be used to transform \code{j} in \verb{[['data']]}.}

\item{cofactor}{Numeric -- default 5000; if \code{transform.type} = \link{asinh} (default), the cofactor will be used to modify the transformation.}
}
\value{
UPDATES BY REFERENCE:
\itemize{
\item \code{flowstate[['data']]}; transformed values
\item \code{flowstate[['parameters']]}; an attribute -- \code{transformed} -- is added to record which parameters have been transformed.
}

Invisibly returns \code{flowstate}.
}
\description{
Linear values in \verb{[['data']]} are transformed using a defined function.  For full spectrum cytometry, transforming linear fluorescence values using \link{asinh} is preferred, with around-zero compression adjusted using a 'cofactor' of 5,000.
}
\examples{
fcs.file.paths <- system.file("extdata", package = "flowstate") |>
list.files(full.names = TRUE, pattern = "BLOCK.*.fcs")

#read .fcs files as a flowstate object; concatenate
fs <- read.flowstate(
  fcs.file.paths,
  colnames.type = "S",
  concatenate = TRUE
)

#plot and mean values of linear columns
plot(fs, CD3, CD8) + ggplot2::guides(fill = 'none')
fs$data[, sapply(.SD, mean), .SDcols = c('CD3', 'CD4', 'CD8')]

#transform
flowstate.transform(
  fs,
  j = c('CD3', 'CD4', 'CD8'),
  transform.type = "asinh",
  cofactor = 5000
)

#updated [['parameters']] attribute
#applied transform function and cofactor mapped to $PnN
attr(fs$parameters,'transformed')

#plot and mean values of transformed columns from updated fs[['data']]
plot(fs, CD3, CD8) + ggplot2::guides(fill = 'none')
fs$data[, sapply(.SD, mean), .SDcols = c('CD3', 'CD4', 'CD8')]

#transform all/remaining
flowstate.transform(
  fs,
  j = NULL,
  transform.type = "asinh",
  cofactor = 5000
)

#updated [['parameters']] attribute
#applied transform function and cofactor mapped to $PnN
attr(fs$parameters,'transformed')

}
