% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_unmixing.R
\name{flowstate.unmix}
\alias{flowstate.unmix}
\title{Unmix a \code{flowstate}}
\usage{
flowstate.unmix(flowstate.object.raw, ref.medians, hash.historic = NULL)
}
\arguments{
\item{flowstate.object.raw}{A \code{flowstate} -- the return of \link{read.flowstate}; the \code{flowstate} must be raw/overdetermined.}

\item{ref.medians}{A \link[data.table:data.table]{data.table} containing normalized \verb{[0,1]} reference control medians -- the return of \link{reference.group.medians}.}

\item{hash.historic}{Character string -- default \code{NULL}; for internal/reproducibility purposes, a hash string of the 'unmixing matrix' can be defined.}
}
\value{
A \code{flowstate} containing unmixed data.
}
\description{
A raw/overdetermined \code{flowstate} is unmixed using a corresponding \code{ref.medians}. Data is unmixed using \link[stats:lsfit]{Ordinary Least Squares}.
}
