% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowstate_object_plot.R
\name{plot.flowstate}
\alias{plot.flowstate}
\title{Plot flowstate data}
\usage{
\method{plot}{flowstate}(x, ..., bins = 200, limits = NULL, sample.n = NULL)
}
\arguments{
\item{x}{A flowstate object as returned from  \link{flowstate}.}

\item{...}{... \link[ggplot2]{aes} arguments (unquoted); essentially x and y. If a \code{z} aesthetic is provided, the plot will switch to \link[ggplot2]{stat_summary_hex}, using this defined third variable as a 'color-by'.}

\item{bins}{\link[ggplot2]{geom_hex} argument; numeric vector giving the number of bins in both vertical and horizontal directions.}

\item{limits}{\link[ggplot2]{continuous_scale} argument.}

\item{sample.n}{Numeric (length 1); if defined, will randomly sample events from \verb{[[data]]}.}
}
\value{
A \link[ggplot2]{ggplot} object.
}
\description{
Plot flowstate data
}
