% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_events.R
\name{plot_spectral.events}
\alias{plot_spectral.events}
\title{Plot Spectral Events}
\usage{
plot_spectral.events(flowstate.object.reference)
}
\arguments{
\item{flowstate.object.reference}{A \code{flowstate} object as returned from \link{reference.group.spectral.events}.}
}
\value{
A list of \link[ggplot2:ggplot]{ggplot} objects.
}
\description{
Plot the results of \link{reference.group.spectral.events}; visual assessment of selected 'spectral events' is used to gauge function performance.
}
