% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_unmixing.R
\name{plot_spectral.trace}
\alias{plot_spectral.trace}
\title{Plot Spectral Traces}
\usage{
plot_spectral.trace(ref.medians)
}
\arguments{
\item{ref.medians}{The return of \link{reference.group.medians}.}
}
\value{
A list of \link[ggplot2:ggplot]{ggplot} objects.
}
\description{
Plots a 'spectral trace' -- normalized \verb{[0,1]} emission for all detectors.
}
