% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_unmixing.R
\name{reference.group.medians}
\alias{reference.group.medians}
\title{Generate normalized medians/spectra from reference group 'spectral events'}
\usage{
reference.group.medians(
  flowstate.object.reference,
  name.fix = NULL,
  syntactically.valid = FALSE
)
}
\arguments{
\item{flowstate.object.reference}{A \code{flowstate} object as returned from \link{reference.group.spectral.events}.}

\item{name.fix}{Named character vector -- default \code{NULL}; if defined, vector names should match to \code{sample.id}(s) and their respective values name replacements.
\itemize{
\item e.g., \code{c('Cd8 BV786 (Cells)' = 'CD8 BV786')}.
}}

\item{syntactically.valid}{Logical -- default \code{FALSE}; if \code{TRUE}, spaces, dashes, and dots are removed from strings.}
}
\value{
A \link[data.table:data.table]{data.table} containing normalized reference control medians.
}
\description{
The 'spectral events' returned from \link{reference.group.spectral.events} are used to generate normalized \verb{[0,1]} medians/spectra for use in spectral unmixing.
}
