% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_group_spectra.R
\name{reference.group.spectra}
\alias{reference.group.spectra}
\title{Title}
\usage{
reference.group.spectra(
  raw.reference.group,
  name.fix = NULL,
  population.marker = NULL,
  top.N.percent = 10,
  scatter.peak.cut = 0.25,
  top.n = 200,
  top.n.override = NULL,
  output.dirs = "derived",
  plot = F
)
}
\arguments{
\item{raw.reference.group}{\link[base]{file.path} -- one of:
\itemize{
\item a reference group directory containing raw single-color/universal negative control .fcs files; all will be processed.
\item a vector of files within that directory; used to process a subset of controls.
}}

\item{name.fix}{Named character vector -- default \code{NULL}; if defined, vector names should match to \code{sample.id}(s) and their respective values name replacements.
\itemize{
\item e.g., \code{c('cd8alpha BV786 (Cells)' = 'CD8a BV786')}.
}}

\item{population.marker}{Named character vector -- default \code{NULL}; \code{population.marker} must be defined for the function to be successful (for cellular controls). The named character vector should take the following form:
\itemize{
\item \code{c(population.name1 = 'sample.name1',population.name2 = 'sample.name2',...)}, where \code{population.name(s)} are CD/lineage cell types and \code{'sample.name(s)'} are CD/lineage markers used to stain those respective cell types.
\itemize{
\item e.g., \verb{c(lymphocytes = 'CD3 BV510 (Cells)', monocytes = 'CD14 SB550 (Cells)'}.
}
}}

\item{top.N.percent}{Numeric -- default \code{10}; defines the percentage of top expressing events (peak detector) used in conjunction with \code{population.marker} for detecting and assigning the named populations.}

\item{scatter.peak.cut}{Numeric -- default \code{0.25}; defines the value at which scatter peak heights will be cut for fine tuning the scatter profile of selected populations (as defined through \code{population.marker}).}

\item{top.n}{Numeric -- default \code{200}; the number of maximally expressing events (sorted vector) used to define 'spectral events' -- events used to define detector medians.}

\item{top.n.override}{Named numeric vector -- default \code{NULL}; if defined, the supplied numeric (value) will override \code{top.n} on a sample-specific (name) basis. See example.}

\item{output.dirs}{Character vector/an output directory -- default \code{derived}; the default will derive output directories based on the location of \code{raw.reference.group} file(s).}

\item{plot}{Logical -- default \code{FALSE}; plots diagnostic/QC output for evaluating function performance.}
}
\value{
A \code{flowstate} containing spectrally-associated bead/cellular events.
}
\description{
This function is entirely dependent -- as of now -- on the following:
\itemize{
\item Spectroflo software raw reference controls  -- tested using SpectroFlo 3.3.0.
\item Spectroflo software naming convention:
\itemize{
\item Directory -- './Raw/Reference Group/...'
\item File names -- 'MARKER FLUOR (Beads|Cells).fcs' (literal space characters separating)
\item Quality reference controls -- well resolved, dominant scatter populations
}
}
}
\examples{
# placeholder
}
