% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowstate_utils.R
\name{select.nonsaturating}
\alias{select.nonsaturating}
\title{Create a column (logical) for selecting against saturating events}
\usage{
select.nonsaturating(flowstate)
}
\arguments{
\item{flowstate}{A flowstate as returned from \link{read.flowstate}.}
}
\value{
UPDATES BY REFERENCE:
\itemize{
\item \code{flowstate[['data']]}; adds a column (logical) named \code{select.nonsaturating}
}

Invisibly returns \code{flowstate}.
}
\description{
A new column (logical) named \code{select.nonsaturating} is created in \verb{[['data']]} and any/all events that are at detector limits -- as defined by the '$PnR' value in \verb{[['parameters']]} -- will be marked as \code{FALSE} with non-saturating events marked as \code{TRUE}.

Detection of saturating events requires linear (non-transformed) values and as such should be performed before \link{flowstate.transform}.
}
\examples{

fcs.file.paths <- system.file("extdata", package = "flowstate") |>
list.files(full.names = TRUE, pattern = "BLOCK.*.fcs")

#read all .fcs files as flowstates; concatenate into a single object
fs <- read.flowstate(
  fcs.file.paths,
  colnames.type = "S",
  concatenate = TRUE
)

#UPDATES BY REFERENCE -- adds a new column named 'select.nonsaturating'
select.nonsaturating(fs)
fs$data[, .N, by = select.nonsaturating]

#visualize
plot(fs,FSC_A,SSC_A) + ggplot2::facet_wrap(~select.nonsaturating)

#subset to retain only non-saturating events
fs <- subset(fs,select.nonsaturating)
fs$data[, .N, by = select.nonsaturating]

#after the subset, the column is now redundant
all(fs$data[['select.nonsaturating']])

#NULL it out
fs$data[,'select.nonsaturating' := NULL]

}
