% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowstate_utils.R
\name{select.quantile}
\alias{select.quantile}
\title{Create a column (logical) for selecting against events below/above a quantile range}
\usage{
select.quantile(flowstate, probs = c(5e-04, 0.9995))
}
\arguments{
\item{flowstate}{A flowstate as returned from \link{read.flowstate}.}

\item{probs}{numeric vector of probabilities with values in
    \eqn{[0,1]}.  (Values up to \samp{2e-14} outside that
    range are accepted and moved to the nearby endpoint.)}
}
\value{
UPDATES BY REFERENCE:
\itemize{
\item \code{flowstate[['data']]}; adds a column (logical) named \code{select.quantile}
}

Invisibly returns \code{flowstate}.
}
\description{
Create a column (logical) for selecting against events below/above a quantile range
}
\examples{

fcs.file.paths <- system.file("extdata", package = "flowstate") |>
list.files(full.names = TRUE, pattern = "BLOCK.*.fcs")

#read all .fcs files as flowstates; concatenate into a single object
fs <- read.flowstate(
  fcs.file.paths,
  colnames.type = "S",
  concatenate = TRUE
)

#transform
flowstate.transform(fs)

#UPDATES BY REFERENCE -- adds a new column named 'select.quantile'
select.quantile(fs, probs = c(0.0005, 0.9995))
fs$data[, .N, by = select.quantile]

#visualize
plot(fs, CD3, CD8) + ggplot2::facet_wrap( ~ select.quantile)

#subset to retain only events in quantile range
fs <- subset(fs, select.quantile)
fs$data[, .N, by = select.quantile]

#after the subset, the column is now redundant
all(fs$data[['select.quantile']])

#NULL it out
fs$data[,'select.quantile' := NULL]

}
