% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_logical.R
\name{select_population}
\alias{select_population}
\title{Create a column (logical) for selecting scatter/population-specific events}
\usage{
select_population(
  flowstate,
  population,
  bandwidth.adjust = 4,
  threshold = 0.75,
  plot = F
)
}
\arguments{
\item{flowstate}{A \code{flowstate} as returned from \link{read.flowstate}.}

\item{population}{Named character vector -- default \code{NULL}; \code{population} MUST be defined. The named character vector should take the following form:
\itemize{
\item \code{c(population.name = 'marker')}, where \code{population.name} is a cell-type annotation and \code{'marker'} is a CD/lineage marker used to stain those respective cell types.
\itemize{
\item e.g., \code{c(lymphocytes = 'CD3')} ;  \code{c(monocytes = 'CD14')}
}
}}

\item{bandwidth.adjust}{Numeric -- default \code{4}; adjusts the 'smoothness' of the \link[MASS:bandwidth.nrd]{density estimation}}

\item{threshold}{Numeric -- default \code{0.75}; fraction of cells/events bound by a contour line.}

\item{plot}{Logical -- default \code{FALSE}; plot the derived contour line.}
}
\value{
UPDATES BY REFERENCE:
\itemize{
\item \code{flowstate[['data']]}:
\itemize{
\item adds a column (logical) named \code{select.population}
\item adds a column (character) named \code{population} -- indexed against \code{select.population}
}
}

Invisibly returns \code{flowstate}.
}
\description{
A new column (logical) named \code{select.population} is created in \verb{[['data']]} along with a respective annotation column (character) named \code{population}. Through the defined \code{population} argument, a specific column in \verb{[['data']]} -- a cluster of differentiation (CD)/lineage surface marker -- is used to derive a scatter/population-specific landmark/profile for restricting cellular events to a population-of-interest.  Classical/specific cell types (e.g., lymphocytes (CD3+), monocytes (CD14+)) can be defined in this fashion.

Based on the the defined \code{population} argument:
\itemize{
\item the top 25\% expressing events in the \code{'marker'} column are indexed;
\item contour lines are derived using \link[MASS:kde2d]{kde2d} on the indexed events;
\item the contour line that meets the \code{threshold} (\% of events in-bounds) is derived;
\item the 'in-bounds' contour line is applied to the entire concatenate to generate both a logical and annotation column
}
}
\examples{
fcs.file.paths <- system.file("extdata", package = "flowstate") |>
list.files(full.names = TRUE, pattern = "BLOCK.*.fcs")

#read all .fcs files as flowstates; concatenate into a single object
fs <- read.flowstate(
  fcs.file.paths,
  colnames.type = "S",
  concatenate = TRUE
)

#transform
flowstate.transform(fs)

#UPDATES BY REFERENCE -- adds two columns: 'select.population' and 'population'
select_population(fs, population = c(lymphocytes = 'CD3'))
fs$data[, .N, by = .(select.population, population)]

#visualize
plot(fs, FSC_A, SSC_A) + ggplot2::facet_wrap( ~ population)

#subset to retain only population-specific events
fs <- subset(fs, population == 'lymphocytes')
fs$data[, .N, by = .(select.population, population)]

#after the subset, the logical column is now redundant; save the annotation
all(fs$data[['select.population']])

#NULL it out
fs$data[,'select.population' := NULL]

}
