% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_group_spectra.R
\name{select_scatter.population}
\alias{select_scatter.population}
\title{Add scatter-specific populations to \verb{[['data']]}}
\usage{
select_scatter.population(
  flowstate,
  population.marker,
  top.N.percent = 10,
  scatter.peak.cut = 0.25,
  plot = F,
  plot.output.dir = tempdir()
)
}
\arguments{
\item{flowstate}{A flowstate object as returned from \link{read.flowstate}.}

\item{population.marker}{Named character vector -- default \code{NULL}; \code{population.marker} must be defined for the function to be successful (for cellular controls). The named character vector should take the following form:
\itemize{
\item \code{c(population.name1 = 'sample.name1',population.name2 = 'sample.name2',...)}, where \code{population.name(s)} are CD/lineage cell types and \code{'sample.name(s)'} are CD/lineage markers used to stain those respective cell types.
\itemize{
\item e.g., \verb{c(lymphocytes = 'CD3 BV510 (Cells)', monocytes = 'CD14 SB550 (Cells)'}.
}
}}

\item{top.N.percent}{Numeric -- default \code{10}; defines the percentage of top expressing events (peak detector) used in conjunction with \code{population.marker} for detecting and assigning the named populations.}

\item{scatter.peak.cut}{Numeric -- default \code{0.25}; defines the value at which scatter peak heights will be cut for fine tuning the scatter profile of selected populations (as defined through \code{population.marker}).}

\item{plot}{Logical -- default \code{FALSE}; plots diagnostic/QC output for evaluating function performance.}

\item{plot.output.dir}{A character vector of a full path name -- default \link[base:tempfile]{tempdir()}; any/all generated plots will be saved to this directory.}
}
\value{
UPDATES BY REFERENCE and invisibly returns \code{flowstate}; a new column (factored) named \code{population} is added to \verb{[['data']]}.
}
\description{
Effective processing of single-color controls first involves defining dominant scatter-specific populations that are free of debris/aggregates and contain max-expressing events.  To efficiently select these populations in cellular controls, a defined set of cluster of differentiation (CD)/lineage surface markers are used to restrict peak-finding/cutting methods to the scatter profile of specific immune cell types.  Once these profiles/populations are defined, they are applied to all of \verb{[['data']]}.
}
\keyword{internal}
