% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_logical.R
\name{select_singlets}
\alias{select_singlets}
\title{Create a column (logical) for selecting singlet (scatter pulse) events}
\usage{
select_singlets(flowstate, quantiles = c(0.85, 0.95))
}
\arguments{
\item{flowstate}{A flowstate as returned from \link{read.flowstate}.}

\item{quantiles}{Numeric vector (length 2) -- default \code{c(0.85, 0.95)}; respective quantile probabilities used to select singlet events first by \verb{FSC-A vs FSC-H}, then by \verb{SSC-A vs SSC-H}.}
}
\value{
UPDATES BY REFERENCE:
\itemize{
\item \code{flowstate[['data']]}; adds a column (logical) named \code{select.singlets}
}

Invisibly returns \code{flowstate}.
}
\description{
A new column (logical) named \code{select.singlets} is created in \verb{[['data']]} and any/all singlet events will be marked as \code{TRUE} with doublet events marked as \code{FALSE}; singlet events are identified through scatter pulse geometry (FSC-A/FSC-H ; SSC-A/SSC-H).
}
\examples{
fcs.file.paths <- system.file("extdata", package = "flowstate") |>
list.files(full.names = TRUE, pattern = "BLOCK.*.fcs")

#read all .fcs files as flowstates; concatenate into a single object
fs <- read.flowstate(
  fcs.file.paths,
  colnames.type = "S",
  concatenate = TRUE
)

#UPDATES BY REFERENCE -- adds a new column named 'select.singlets'
select_singlets(fs, quantiles = c(0.95, 0.95))
fs$data[, .N, by = select.singlets]

#visualize
plot(fs, FSC_A, FSC_H) + ggplot2::facet_wrap( ~ select.singlets)
plot(fs, SSC_A, SSC_H) + ggplot2::facet_wrap( ~ select.singlets)

#subset to retain only singlet events
fs <- subset(fs, select.singlets)
fs$data[, .N, by = select.singlets]

#after the subset, the column is now redundant
all(fs$data[['select.singlets']])

#NULL it out
fs$data[,'select.singlets' := NULL]

}
