% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spillover_flowstate.R
\name{spillover.apply}
\alias{spillover.apply}
\title{Compensate \code{flowstate[['data']]} using values stored in \code{flowstate[['spill']]}}
\usage{
spillover.apply(flowstate, decompensate = FALSE)
}
\arguments{
\item{flowstate}{A flowstate object as returned from \link{read.flowstate}.}

\item{decompensate}{Logical; if \code{TRUE}, will return data values to their original state.}
}
\value{
UPDATES BY REFERENCE:
\itemize{
\item \code{flowstate[['data']]}; applies compensation to data using values stored in spill.
}

Invisibly returns \code{flowstate}.
}
\description{
Compensate \code{flowstate[['data']]} using values stored in \code{flowstate[['spill']]}
}
\examples{
fcs.file.paths <- system.file("extdata", package = "flowstate") |>
list.files(full.names = TRUE, pattern = "BLOCK.*.fcs")

#read all .fcs files as flowstates; concatenate into a single object
fs <- read.flowstate(
  fcs.file.paths,
  colnames.type = "S",
  concatenate = TRUE
)

#transform
flowstate.transform(fs, c('CD4', 'CD8'))

#row index; CD4 vs CD8
index <- which(names(fs$spill) \%in\% c('CD4', 'CD8'))

#update a spill value
spillover.update.value(fs, CD8, CD4, 0.03)

#over-compensated data
plot(fs, CD4, CD8) + ggplot2::labs(title = "Over-compensated")

#apply compensation
spillover.apply(fs)

#compensated data
plot(fs, CD4, CD8) + ggplot2::labs(title = "Compensated")

#return data values to original state by decompensating
spillover.apply(fs, decompensate = TRUE)

plot(fs, CD4, CD8) + ggplot2::labs(title = "Over-compensated")

}
