% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spillover_flowstate.R
\name{spillover.update.value}
\alias{spillover.update.value}
\title{Update the values of \code{flowstate[['spill']]}}
\usage{
spillover.update.value(flowstate, i, j, value)
}
\arguments{
\item{flowstate}{A flowstate object as returned from \link{read.flowstate}.}

\item{i}{Variable (unquoted).}

\item{j}{Variable (unquoted).}

\item{value}{Numeric; correction value to be used during compensation.}
}
\value{
UPDATES BY REFERENCE:
\itemize{
\item \code{flowstate[['spill']]}; updates \verb{[i, j]} with the defined correction value.
}

Invisibly returns \code{flowstate}.
}
\description{
Update the values of \code{flowstate[['spill']]}
}
\examples{
fcs.file.paths <- system.file("extdata", package = "flowstate") |>
list.files(full.names = TRUE, pattern = "BLOCK.*.fcs")

#read all .fcs files as flowstate objects; concatenate into a single object
fs <- read.flowstate(
  fcs.file.paths,
  colnames.type = "S",
  concatenate = TRUE
)

#row index; CD4 vs CD8
index <- which(names(fs$spill) \%in\% c('CD4', 'CD8'))
fs$spill[index, .(CD4, CD8)]

#update a spill value
spillover.update.value(fs, i = CD8, j = CD4, value = 0.03)

fs$spill[index, .(CD4, CD8)]

}
