% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowstate_S3.R
\name{split.flowstate}
\alias{split.flowstate}
\title{Split a concatenated \code{flowstate} object into a list}
\usage{
\method{split}{flowstate}(x, f, drop = TRUE, ...)
}
\arguments{
\item{x}{A concatenated flowstate object as returned from \link{read.flowstate}.}

\item{f}{Character string; a single (factored) column by which \verb{[['data']]} will be split into individual \code{flowstate}s.}

\item{drop}{Logical; default \code{TRUE}. Unused factor levels in \code{f} are dropped -- no \code{flowstate} object will be returned.}

\item{...}{Potential further arguments.}
}
\value{
A list of individual \code{flowstate}s.
}
\description{
Split a concatenated \code{flowstate} object into a list
}
\examples{
fcs.file.paths <- system.file("extdata", package = "flowstate") |>
list.files(full.names = TRUE, pattern = "BLOCK.*.fcs")

#read all .fcs files as flowstate objects; concatenate into a single object
fs <- read.flowstate(
  fcs.file.paths,
  colnames.type="S",
  concatenate = TRUE
)

fs.split <- split(fs,'sample.id')

length(fs.split)
names(fs.split)
lapply(fs.split,'[[','keywords')

#remove all rows associated with sample 1; level will remain
fs$data <- fs$data[!sample.id \%in\% levels(sample.id)[1]]

#with drop
fs.split <- split(fs,'sample.id',drop = TRUE)
length(fs.split)
names(fs.split)

#without drop
fs.split <- split(fs,'sample.id',drop = FALSE)
length(fs.split)
names(fs.split)
#empty data for sample 1
fs.split[[1]]$data

}
