% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowstate_S3.R
\name{subset.flowstate}
\alias{subset.flowstate}
\title{Subset \code{flowstate[['data']]}}
\usage{
\method{subset}{flowstate}(x, ...)
}
\arguments{
\item{x}{A flowstate object as returned from \link{read.flowstate}.}

\item{...}{\code{subset} argument as defined by \code{data.table}'s \link[data.table]{subset}; logical expression indicating elements or rows to keep.}
}
\value{
A \code{flowstate} object whose \verb{[['data']]} element (a \code{data.table}) has been subset to contain the rows as defined through the \code{subset} argument. Other elements of \code{x} are copied.
}
\description{
Subset \code{flowstate[['data']]} and return a new \code{flowstate} object.
}
\examples{
fcs.file.paths <- system.file("extdata", package = "flowstate") |>
list.files(full.names = TRUE, pattern = "BLOCK.*.fcs")

#read all .fcs files as flowstate objects; concatenate into a single object
fs <- read.flowstate(
  fcs.file.paths,
  colnames.type="S",
  concatenate = TRUE
)

#transform
flowstate.transform(fs,'CD3')

#plot;visualize valley
fs$data[,plot(density(CD3))]
abline(v = 1.1)#not data-driven; just for example

#subset
fs.cd3_positive <- subset(fs,CD3 > 1.1)

#totals
fs$data[,.N]
fs.cd3_positive$data[,.N]
}
