% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_flowstate.R
\name{write.flowstate}
\alias{write.flowstate}
\title{Write a \code{flowstate} as a Flow Cytometry Standard file (FCS 3.1)}
\usage{
write.flowstate(
  flowstate,
  file.dir = tempdir(),
  fil.suffix = NULL,
  endianness = c("little", "big")
)
}
\arguments{
\item{flowstate}{A \code{flowstate} as returned from \link{read.flowstate}.}

\item{file.dir}{A \link{file.path} -- default \link{tempdir}; a file directory for saving the output FCS 3.1 file(s).}

\item{fil.suffix}{Character string -- default \code{NULL}; if defined, the suffix will be appended}

\item{endianness}{Character string -- default \code{"little"}; see \link[base:writeBin]{endian}.}
}
\value{
\itemize{
\item FCS 3.1 file(s) are written to disk and a summary message is displayed.
\itemize{
\item the original \verb{$FIL} will be appended with \code{'flowstateMOD'} as an explicit indicator of modification; if \code{fil.suffix} is defined, \code{'flowstateMOD'} will be further appended with the string value.
}
\item UPDATES BY REFERENCE:
\itemize{
\item \code{flowstate[['data']]} -- non-numeric (logical and/or factor) columns are converted to numeric
\item \code{flowstate[['keywords']]} -- non-character (factor) columns are converted to character
\item \code{flowstate[['spill']]} -- column names are matched/updated using \verb{[['parameters']][['N']]}
}
\item UPDATES BY ASSIGNMENT to environment:
\itemize{
\item \code{flowstate[['parameters']]} -- \verb{[['parameters']]} is updated by row-binding derived columns found in \verb{[['data']]}
}
}

Invisibly returns \code{flowstate}.
}
\description{
A \code{flowstate} -- following \href{https://pmc.ncbi.nlm.nih.gov/articles/PMC2892967/}{Flow Cytometry Standard} conventions -- is converted to FCS 3.1 and written to disk. The conversion is as follows:
\itemize{
\item \verb{[['keywords']]}, \verb{[['parameters']]}, and \verb{[['spill']]} (if present) are converted to character string using a delimiter \code{"|"} to separate keyword-value pairs.
\itemize{
\item this string forms the \code{TEXT} segment of the FCS file.
}
\item \verb{[['data']]} is written as row-major binary.
\itemize{
\item this forms the \code{DATA} segment of the FCS file.
}
\item The \code{HEADER} segment (string) -- containing the required offsets -- is derived from \code{TEXT} and \code{DATA}.
}

The FCS 3.1 compliant file is then written to disk.

\emph{N.B.: During testing, FCS 3.1 files written to disk by \link{write.flowstate} were fully compatible with \link{read.flowstate} and \link[flowCore:read.FCS]{flowCore::read.FCS}; issues were only with proprietary software (i.e., FlowJo). If proprietary software is to be used with FCS 3.1 files generated by \link{write.flowstate}, caution is advised.}
}
\examples{

fcs.file.paths <- system.file("extdata", package = "flowstate") |>
list.files(full.names = TRUE, pattern = "BLOCK.*.fcs")

## read .fcs files as a flowstate object; concatenate
fs <- read.flowstate(
  fcs.file.paths,
  colnames.type = "S",
  concatenate = TRUE
)

## saturating events -- logical
select_nonsaturating(fs)

## singlet events -- logical
select_singlets(fs)

## transform
flowstate.transform(fs)

## simple logical to select CD3+
fs$data[
,
j = select.CD3 := {
 d <- stats::density(CD3[(select.nonsaturating) & (select.singlets)])
 pv <- diff(sign(diff(d$y)))
 valley <- which(pv == 2) + 1
 CD3 > d$x[valley]
 }
]

## add a factor -- 'aliquot'
fs$keywords[
 ,
 j = aliquot := data.table::tstrsplit(sample.id, "_", keep = 5, type.convert = factor)
]

add.keywords.to.data(fs, 'aliquot')

## write; reassign to capture the updated flowstate
fs <- write.flowstate(fs, fil.suffix = "Example")

## read newly written files
fcs.file.paths <- list.files(
 tempdir(),
 full.names = TRUE,
 pattern = "flowstateMOD_Example.fcs"
)

fs <- read.flowstate(
  fcs.file.paths,
  colnames.type = "S",
  concatenate = TRUE
)

## derived [['data']] columns
fs[['data']][, .(select.nonsaturating, select.singlets, select.CD3, aliquot)]

## derived [['parameters']] columns
fs$parameters[grep("Logical|Factor", TYPE)]

## derived [['keyword']] columns -- retains original factor levels
fs[['keywords']][, .(aliquot)]
## recover the factor levels -- optional
levels.aliquot <- unique(fs[['keywords']][, sub('FACTOR::', "", aliquot)])
levels.aliquot <- strsplit(levels.aliquot, ';')[[1]]
fs$keywords[, aliquot := factor(levels.aliquot)]
fs$data[, aliquot := NULL]
add.keywords.to.data(fs, 'aliquot')

## plot derived [['data']] logicals (now numeric)
plot(fs, FSC_A, FSC_H) + ggplot2::facet_wrap(~select.singlets)
flowstate.transform(fs)
plot(fs, CD3) + ggplot2::facet_wrap(~select.CD3)

}
