% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concatenate_flowstate.R
\name{concatenate.flowstate}
\alias{concatenate.flowstate}
\title{Concatenate a list of flowstate objects}
\usage{
concatenate.flowstate(flowstate.objects)
}
\arguments{
\item{flowstate.objects}{a list; the return of \link{read.flowstate}}
}
\value{
An object of class flowstate.
}
\description{
Concatenate a list of flowstate objects
}
\examples{
fcs.file.paths <- system.file("extdata", package = "flowstate") |>
list.files(full.names = TRUE, pattern = "BLOCK.*.fcs")

#read all .fcs files as flowstate objects
fs <- read.flowstate(
  fcs.file.paths,
  colnames.type="S"
)

#a list of flowstate objects
class(fs)
sapply(fs,class)

#concatenate into a single flowstate object
fs<-concatenate.flowstate(fs)
class(fs)

fs$data
fs$parameters
fs$keywords
}
