% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_flowstate.R
\name{flowstate.transform.inverse}
\alias{flowstate.transform.inverse}
\title{Transform \code{flowstate[['data']]} -- Inverse}
\usage{
flowstate.transform.inverse(flowstate.object, .j = NULL)
}
\arguments{
\item{flowstate.object}{A flowstate object as returned from \link{read.flowstate}.}

\item{.j}{Character vector -- default \code{NULL}; any/all parameters having a keyword-value pair of \code{transform/asinh} will be inverse transformed in \verb{[['data']]} using \code{\link[base:Hyperbolic]{base::sinh()}}.  If a character vector: specific columns in \verb{[['data']]} that are to be inverse transformed.}
}
\value{
UPDATES BY REFERENCE:
\itemize{
\item \code{flowstate[['data']]}; inverse transformed values -- linear
\item \code{flowstate[['parameters']]}; modifies two columns ('transform' and 'cofactor') -- sets to \code{NA}
}

Invisibly returns the \code{flowstate.object}.
}
\description{
Transform \code{flowstate[['data']]} -- Inverse
}
\examples{
fcs.file.paths <- system.file("extdata", package = "flowstate") |>
list.files(full.names = TRUE, pattern = "BLOCK.*.fcs")

#read .fcs files as a flowstate object; concatenate
fs <- read.flowstate(
  fcs.file.paths,
  colnames.type="S",
  concatenate = TRUE
)

#plot and mean values of linear columns
plot(fs,CD3,CD8) + ggplot2::guides(fill = 'none')
res1.linear <- fs$data[,sapply(.SD,mean),.SDcols = c('CD3','CD8')]
print(res1.linear)

#transform
flowstate.transform(
  fs,
  .j = c('CD3','CD8'),
  transform.type = "asinh",
  cofactor = 5000
)
#updated parameters
fs$parameters[!is.na(transform)]

#plot and mean values of transformed columns from updated fs[['data']]
plot(fs,CD3,CD8) + ggplot2::guides(fill = 'none')
fs$data[,sapply(.SD,mean),.SDcols = c('CD3','CD8')]

#inverse transformation
flowstate.transform.inverse(fs)

#updated parameters; transform and cofactor set to NA
fs$parameters[S \%in\% c('CD3','CD8')]

#plot and mean values of linear columns
plot(fs,CD3,CD8) + ggplot2::guides(fill = 'none')
res2.linear <- fs$data[,sapply(.SD,mean),.SDcols = c('CD3','CD8')]
print(res2.linear)

#linear --> transformed --> inverse --> linear
res1.linear == res2.linear

}
