% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_flowstate.R
\name{flowstate.transform}
\alias{flowstate.transform}
\title{Transform \code{flowstate[['data']]}}
\usage{
flowstate.transform(
  flowstate.object,
  .j = NULL,
  transform.type = "asinh",
  cofactor = 5000
)
}
\arguments{
\item{flowstate.object}{A flowstate object as returned from \link{read.flowstate}.}

\item{.j}{Character vector -- default \code{NULL}; any/all parameters having a keyword-value pair of \code{TYPE/Raw|Unmixed_Fluorescence} will be transformed in \verb{[['data']]}.  If a character vector: specific columns in \verb{[['data']]} that are to be transformed.}

\item{transform.type}{Character string -- default \link{asinh}; quoted function that will be used to transform \code{.j} in \verb{[['data']]}.}

\item{cofactor}{Numeric -- default 5000; if \code{transform.type} = \link{asinh} (default), the cofactor will be used to modify the transformation.}
}
\value{
UPDATES BY REFERENCE:
\itemize{
\item \code{flowstate[['data']]}; transformed values
\item \code{flowstate[['parameters']]}; adds two columns ('transform' and 'cofactor')
}

Invisibly returns the \code{flowstate.object}.
}
\description{
Transform \code{flowstate[['data']]}
}
\examples{
fcs.file.paths <- system.file("extdata", package = "flowstate") |>
list.files(full.names = TRUE, pattern = "BLOCK.*.fcs")

#read .fcs files as a flowstate object; concatenate
fs <- read.flowstate(
  fcs.file.paths,
  colnames.type="S",
  concatenate = TRUE
)

#plot and mean values of linear columns
plot(fs,CD3,CD8) + ggplot2::guides(fill = 'none')
fs$data[,sapply(.SD,mean),.SDcols = c('CD3','CD4','CD8')]

#transform
flowstate.transform(
  fs,
  .j = c('CD3','CD4','CD8'),
  transform.type = "asinh",
  cofactor = 5000
)
#updated parameters
fs$parameters[!is.na(transform)]

#plot and mean values of transformed columns from updated fs[['data']]
plot(fs,CD3,CD8) + ggplot2::guides(fill = 'none')
fs$data[,sapply(.SD,mean),.SDcols = c('CD3','CD4','CD8')]

#transform all
flowstate.transform(
  fs,
  .j = NULL,
  transform.type = "asinh",
  cofactor = 5000
)

#updated parameters
fs$parameters[!is.na(transform)]

}
