% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_flowstate.R
\name{plot.flowstate}
\alias{plot.flowstate}
\title{Plot flowstate data}
\usage{
\method{plot}{flowstate}(x, ..., bins = 200, limits = NULL, sample.n = NULL)
}
\arguments{
\item{x}{A flowstate object as returned from \link{read.flowstate}.}

\item{...}{... \link[ggplot2]{aes} arguments (unquoted variables); essentially x and y. If a \code{z} aesthetic is defined, the plot will switch to \link[ggplot2]{stat_summary_hex}, using this defined third variable as a 'color-by'.}

\item{bins}{\link[ggplot2]{geom_hex} argument; numeric giving the number of bins in both vertical and horizontal directions.}

\item{limits}{\link[ggplot2]{continuous_scale} argument.}

\item{sample.n}{Numeric (length 1); if defined, will randomly sample events from \verb{[[data]]}.}
}
\value{
A \link[ggplot2]{ggplot} object.
}
\description{
Plot flowstate data
}
\examples{
fcs.file.paths <- system.file("extdata", package = "flowstate") |>
list.files(full.names = TRUE, pattern = "BLOCK.*.fcs")

#read all .fcs files as flowstate objects; concatenate into a single object
fs <- read.flowstate(
  fcs.file.paths,
  colnames.type="S",
  concatenate = TRUE
)

#transform
flowstate.transform(fs,c('CD3','CD4','CD8','Viability'))

#plot title
no.fill.legend <- ggplot2::guides(fill = 'none')
.title1 <- paste("Batch:",fs$keywords[,unique(`$PROJ`)])
.title2 <- paste(
  "Instrument Serial#:",
  fs$keywords[,paste(.(unique(`$CYT`),unique(`$CYTSN`)),collapse = " ")]
)
.title <- paste(.title1,.title2,sep = "\n")
.title <- ggplot2::labs(title = .title)

#plot: two variables
plot(fs,CD3,Viability) + no.fill.legend + .title
plot(fs,FSC_A,SSC_A) + no.fill.legend + .title

#plot: two variables; facet by sample.id
plot(fs,CD4,CD8) + no.fill.legend + .title + ggplot2::facet_wrap(~sample.id)

#plot: three variables; color-by
plot(fs,FSC_A,SSC_A,z=Viability,bins=100,limits=c(0,3)) + .title

}
