% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_flowstate.R
\name{read.flowstate}
\alias{read.flowstate}
\title{\code{flowstate}: read, parse, and store .fcs data}
\usage{
read.flowstate(
  fcs.file.paths,
  colnames.type = c("S_N", "S", "N"),
  S.func = NULL,
  sample.id = NULL,
  concatenate = FALSE
)
}
\arguments{
\item{fcs.file.paths}{Character string; path(s) returned from \code{list.files(...,full.names=T,pattern=".fcs")}.}

\item{colnames.type}{Character string; one of:
\itemize{
\item \code{"S_N"} -- \verb{[['data']]} columns are named by combining $PS (stain) and $PN (name), separated by an underscore.
\item \code{"S"} -- \verb{[['data']]} columns are named by using only their respective $PS (stain) keyword value.
\item \code{"N"} -- \verb{[['data']]} columns are named by using only their respective $PN (name) keyword value.
}}

\item{S.func}{A function -- default \code{NULL}; if a function is supplied, it will be used to modify/split \code{"S"}; e.g. \code{function(j){strsplit(j," ")[[1]][1]}} will be applied to \code{"S"} to return the first split element ("CD4 PE" --> "CD4").}

\item{sample.id}{Keyword name -- default \code{NULL}; based on cytometer platform, the keyword name for \code{sample.id} will be automatically set and the respective keyword values will be added to \verb{[['data']]} as a factored sample identifier. One of:
\itemize{
\item Aurora (Cytek) -- \code{sample.id} = \code{'TUBENAME'}
\item ID7000 (Sony)  -- \code{sample.id} = \code{'$CELLS'}
\item Unspecified    -- \code{sample.id} = \code{'$FIL'}
}}

\item{concatenate}{Logical -- default \code{FALSE}; if \code{TRUE}, the list of flowstate objects will be combined into a single flowstate object.}
}
\value{
For a single file: an object of class \code{flowstate}; for multiple files: a named list of \verb{flowstate objects}; for concatenated files: an object of class \code{flowstate}
}
\description{
\href{https://pmc.ncbi.nlm.nih.gov/articles/PMC2892967/}{Flow Cytometry Standard} files are read, parsed and stored as objects (S3) of \link{class} \code{'flowstate'}.

The individual segments (\code{TEXT} and \code{DATA}) of the FCS file are parsed and stored as follows:
\itemize{
\item \verb{[['data']]} -- a \link[data.table:data.table]{data.table} containing raw/linear measurement values (scatter, MFI, Time, etc.).
\item \verb{[['parameters']]} -- a \link[data.table:data.table]{data.table} containing instrument-specific parameters ('$PnN','$PnS', etc.).
\item \verb{[['keywords']]} -- a \link[data.table:data.table]{data.table} containing instrument/sample-specific keyword-value pairs (metadata).
\item \verb{[['spill']]} -- a \link[data.table:data.table]{data.table} containing (if present) spillover values.
}
}
\details{
Access the individual named list elements as follows (assuming \code{flowstate} object is named \code{fs}):
\itemize{
\item \code{fs$data} or \code{fs[['data']]}
\item \code{fs$parameters} or \code{fs[['parameters']]}
\item \code{fs$keywords} or \code{fs[['keywords']]}
\item \code{fs$spill} or \code{fs[['spill']]}
}

Other \code{flowstate} functions operate on the entire object (\code{fs}) and access the individual slots as needed.
}
\examples{
fcs.file.paths <- system.file("extdata", package = "flowstate") |>
list.files(full.names = TRUE, pattern = "BLOCK.*.fcs")

#read a single .fcs file as a flowstate object
fs <- read.flowstate(
  fcs.file.paths[1],
  colnames.type="S"
)
class(fs)
names(fs)

#.fcs DATA segment as a data.table
  fs$data
#.fcs TEXT segment parsed and stored as three elements (data.tables):
  fs$parameters #instrument-specific measurement parameters
  fs$keywords #instrument/sample-specific metadata
  fs$spill #instrument/sample-specific spillover

#read all .fcs files as a named list containing individial flowstate objects
fs <- read.flowstate(
  fcs.file.paths,
  colnames.type="S",
  concatenate = FALSE
)
class(fs);class(fs[[1]])
names(fs);names(fs[[1]])

fs[[1]]$keywords
fs[[1]]$data[,levels(sample.id)]

#read all .fcs files as flowstate objects; concatenate into a single object
fs <- read.flowstate(
  fcs.file.paths,
  colnames.type="S",
  concatenate = TRUE
)
class(fs)
names(fs)
fs$keywords
fs$data[,levels(sample.id)]

}
\references{
Directly/heavily-inspired by:

\link[flowCore:flowFrame-class]{flowCore}:

Ellis B, Haaland P, Hahne F, Le Meur N, Gopalakrishnan N, Spidlen J, Jiang M, Finak G (2024). \emph{flowCore: flowCore: Basic
structures for flow cytometry data}. doi:10.18129/B9.bioc.flowCore \url{https://doi.org/10.18129/B9.bioc.flowCore}, R package
version 2.22.0, \url{https://bioconductor.org/packages/flowCore}.

\link[data.table:data.table]{data.table}:

Barrett T, Dowle M, Srinivasan A, Gorecki J, Chirico M, Hocking T, Schwendinger B, Krylov I (2025).
data.table: Extension of 'data.frame'. R package version 1.17.99, https://r-datatable.com.
}
\seealso{
\code{\link[=flowstate.transform]{flowstate.transform()}}
}
