% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_flowstate.R
\name{read.flowstate}
\alias{read.flowstate}
\title{flowstate: read, process, and store .fcs data}
\usage{
read.flowstate(
  fcs.file.paths,
  colnames.type = c("S_N", "S", "N"),
  cofactor = 5000,
  sample.id = "TUBENAME",
  concatenate = FALSE
)
}
\arguments{
\item{fcs.file.paths}{Character string; path(s) returned from \code{list.files(...,full.names=T,pattern=".fcs")}.}

\item{colnames.type}{Character string; one of:
\itemize{
\item \code{"S_N"} -- \verb{[['data']]} columns are named by combining $PS (stain) and $PN (name), separated by an underscore.
\item \code{"S"} -- \verb{[['data']]} columns are named by using only their respective $PS (stain) keyword value.
\item \code{"N"} -- \verb{[['data']]} columns are named by using only their respective $PN (name) keyword value.
}}

\item{cofactor}{Numeric; default \code{5000}. Any/all parameters with a \verb{$PnTYPE} of 'Raw_Fluorescence' or 'Unmixed_Fluorescence' will be transformed using \link{asinh} and the defined cofactor value (\code{asinh(x/cofactor)}).}

\item{sample.id}{Character string; the keyword label defined through \code{sample.id} (default \code{TUBENAME}) will be used to add respective keyword values from \verb{[['keywords']]} as an identifier to \verb{[['data']]}.}

\item{concatenate}{Logical; if \code{TRUE}, the list of flowstate objects will be combined into a single flowstate object.}
}
\value{
For a single file: an object of class flowstate; for multiple files: a list of flowstate objects.
}
\description{
flowstate: read, process, and store .fcs data
}
\examples{
fcs.file.paths <- system.file("extdata", package = "flowstate") |>
list.files(full.names = TRUE, pattern = ".fcs")

#read a single .fcs file as a flowstate object
fs <- read.flowstate(
  fcs.file.paths[1],
  colnames.type="S",
  cofactor = 5000
)
class(fs)

#.fcs DATA segment as a data.table
  fs$data
#.fcs TEXT segment parsed and stored as three elements (data.tables):
  fs$parameters #instrument-specific measurement parameters
  fs$keywords #instrument/sample-specific metadata
  fs$spill #instrument/sample-specific spillover

#read all .fcs files as flowstate objects; concatenate into a single object
fs <- read.flowstate(
  fcs.file.paths,
  colnames.type="S",
  cofactor = 5000,
  concatenate = TRUE
)
class(fs)
fs$keywords

}
