% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_flowstate.R
\name{read.flowstate}
\alias{read.flowstate}
\title{flowstate: read, process, and store .fcs data}
\usage{
read.flowstate(
  fcs.file.paths,
  colnames.type = c("S_N", "S", "N"),
  S.func = NULL,
  sample.id = NULL,
  concatenate = FALSE
)
}
\arguments{
\item{fcs.file.paths}{Character string; path(s) returned from \code{list.files(...,full.names=T,pattern=".fcs")}.}

\item{colnames.type}{Character string; one of:
\itemize{
\item \code{"S_N"} -- \verb{[['data']]} columns are named by combining $PS (stain) and $PN (name), separated by an underscore.
\item \code{"S"} -- \verb{[['data']]} columns are named by using only their respective $PS (stain) keyword value.
\item \code{"N"} -- \verb{[['data']]} columns are named by using only their respective $PN (name) keyword value.
}}

\item{S.func}{A function -- default \code{NULL}; if a function is supplied, it will be used to modify/split \code{"S"}; e.g. \code{function(j){strsplit(j," ")[[1]][1]}} will be applied to \code{"S"} to return the first split element ("CD4 PE" --> "CD4").}

\item{sample.id}{Keyword name -- default \code{NULL}; based on cytometer platform, the keyword name for \code{sample.id} will be automatically set and the respective keyword values will be added to \verb{[['data']]} as a factored sample identifier. One of:
\itemize{
\item Aurora (Cytek) -- \code{sample.id} = \code{'TUBENAME'}
\item ID7000 (Sony)  -- \code{sample.id} = \code{'$CELLS'}
\item Unspecified    -- \code{sample.id} = \code{'$FIL'}
}}

\item{concatenate}{Logical -- default \code{FALSE}; if \code{TRUE}, the list of flowstate objects will be combined into a single flowstate object.}
}
\value{
For a single file: an object of class \code{flowstate}; for multiple files: a named list of \verb{flowstate objects}; for concatenated files: an object of class \code{flowstate}
}
\description{
flowstate: read, process, and store .fcs data
}
\examples{
fcs.file.paths <- system.file("extdata", package = "flowstate") |>
list.files(full.names = TRUE, pattern = "BLOCK.*.fcs")

#read a single .fcs file as a flowstate object
fs <- read.flowstate(
  fcs.file.paths[1],
  colnames.type="S"
)
class(fs)
names(fs)

#.fcs DATA segment as a data.table
  fs$data
#.fcs TEXT segment parsed and stored as three elements (data.tables):
  fs$parameters #instrument-specific measurement parameters
  fs$keywords #instrument/sample-specific metadata
  fs$spill #instrument/sample-specific spillover

#read all .fcs files as a named list containing individial flowstate objects
fs <- read.flowstate(
  fcs.file.paths,
  colnames.type="S",
  concatenate = FALSE
)
class(fs);class(fs[[1]])
names(fs);names(fs[[1]])

fs[[1]]$keywords
fs[[1]]$data[,levels(sample.id)]

#read all .fcs files as flowstate objects; concatenate into a single object
fs <- read.flowstate(
  fcs.file.paths,
  colnames.type="S",
  concatenate = TRUE
)
class(fs)
names(fs)
fs$keywords
fs$data[,levels(sample.id)]

}
\seealso{
\code{\link[=flowstate.transform]{flowstate.transform()}}
}
