% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_unmixing.R
\name{reference.group.spectral.events}
\alias{reference.group.spectral.events}
\title{A flowstate containing spectrally-associated bead/cellular events.}
\usage{
reference.group.spectral.events(
  raw.reference.group.directory,
  internal.negative = NULL,
  n = 500,
  plot.select = FALSE,
  plot.select.dir = tempdir()
)
}
\arguments{
\item{raw.reference.group.directory}{Character vector; a file path to a SpectroFlo Raw Reference Group directory.}

\item{internal.negative}{Character vector; spectral events associated with an internal negative population will be generated for any named reference controls.}

\item{n}{Numeric; default 500. Orders the top \code{n} expressing events and aids in auto-detection of peak detector; used by an internal helper function.}

\item{plot.select}{Logical; default \code{FALSE}. Plots diagnostic/QC output for evaluating function performance.}

\item{plot.select.dir}{Character vector; a file path for storing plot output.}
}
\value{
A flowstate containing spectrally-associated bead/cellular events. \verb{[['data']]} will contain appended columns/metadata.
}
\description{
This function is entirely dependent on the following:
\itemize{
\item Spectroflo software raw reference controls  -- tested using SpectroFlo 3.3.0.
\item Spectroflo software naming convention:
\itemize{
\item Directory -- './Raw/Reference Group/...'
\item File names -- 'MARKER FLUOR (Beads|Cells).fcs' (literal space characters separating)
\item Quality reference controls -- well resolved, dominant scatter populations
}
}
}
\examples{
raw.ref.dir <- system.file("extdata", package = "flowstate") |>
list.dirs() |>
grep('Reference Group', x = _, value = TRUE)

ref <- reference.group.spectral.events(
raw.reference.group.directory = raw.ref.dir,
internal.negative = NULL,
n = 500,
plot.select = TRUE,
plot.select.dir = tempdir()
)

#plot output
list.files(tempdir(),pattern = "select.*.pdf")

#spectral events
ref$data

}
