% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_events.R
\name{reference.group.spectral.events}
\alias{reference.group.spectral.events}
\title{A \code{flowstate} containing spectrally-associated bead/cellular events.}
\usage{
reference.group.spectral.events(
  raw.reference.group,
  transform = FALSE,
  quantiles = NULL,
  population.markers = NULL,
  top.N.percent = 1,
  height.cut.scatter = 0.5,
  n.detector.events = 200,
  detector.override = NULL,
  n.spectral.events = 200,
  n.spectral.events.unstained = 1000,
  n.spectral.events.override = NULL,
  plot = FALSE,
  plot.dir = "$PROJ",
  verbose = FALSE
)
}
\arguments{
\item{raw.reference.group}{\link[base]{file.path} -- one of:
\itemize{
\item a reference group directory containing raw single-color/universal negative control .fcs files; all will be processed.
\item a vector of files within that directory; used to process a subset of controls.
}}

\item{transform}{Logical -- default \code{FALSE}; if \code{TRUE}, parameters (detectors) having a keyword-value pair of  \code{TYPE/Raw_Fluorescence} will be transformed using \link[base:Hyperbolic]{asinh(x/5000)}.}

\item{quantiles}{Named \link[base]{list} -- default \code{NULL}; if defined, the list must conform to the following:
\itemize{
\item \code{list('scatter' = probs,'fluor' = probs)}, where \link[stats:quantile]{probs} defines a numeric vector of length 2.
\itemize{
\item e.g., \code{list('scatter' = c(0,0.99),'fluor' = c(0,0.999))}; as defined, 'scatter' (forward and side) and 'fluor' (fluorescence) associated events will be excluded from processing based on their respective \code{probs}.
}
}}

\item{population.markers}{Named character vector -- default \code{NULL}; \code{population.markers} must be defined for the function to be successful (for cellular controls). The named character vector should take the following form:
\itemize{
\item \code{c(population.name1 = 'sample.name1',population.name2 = 'sample.name2',...)}, where \code{population.name(s)} are lineage cell types and \code{'sample.name(s)'} are lineage markers used to stain those respective cell types.
\itemize{
\item e.g., \verb{c(lymphocytes = 'CD3 BV510 (Cells)', monocytes = 'CD14 SB550 (Cells)'}.
}
}
The defined \code{population.markers} will be used as 'cellular anchors': top expressing events (peak detector) in the named sample will be used to 'anchor' scatter distributions for proper assignment of the named population.}

\item{top.N.percent}{Numeric -- default \code{1}; defines the number of top expressing events (peak detector) used in conjunction with \code{population.markers} for detecting and assigning the named populations.}

\item{height.cut.scatter}{Numeric -- default \code{0.5}; defines the value at which scatter peak heights will be cut for fine tuning selected populations (as defined through \code{population.markers}).}

\item{n.detector.events}{Numeric -- default \code{100}; the number of maximally expressing events (sorted vector) used to auto-detect reference control-specific peak detectors.}

\item{detector.override}{Named character vector -- default \code{NULL}; if defined, the supplied detector name (value) will override the auto-detected peak detector on a reference control-specific (name) basis. See example.}

\item{n.spectral.events}{Numeric -- default \code{250}; the number of maximally expressing events (sorted vector) used to define 'spectral events'.}

\item{n.spectral.events.unstained}{Numeric -- default \code{1000}; the number of maximally expressing events (sorted vector) used to define '(universal) negative' events.}

\item{n.spectral.events.override}{Named numeric vector -- default \code{NULL}; if defined, the supplied numeric (value) will override \code{n.spectral.events} on a sample-specific (name) basis. See example.}

\item{plot}{Logical -- default \code{FALSE}; plots diagnostic/QC output for evaluating function performance.}

\item{plot.dir}{Character vector -- default \code{'$PROJ'}; the value associated with the keyword '$PROJ' (experiment name) will be used to construct a plot output directory.}

\item{verbose}{Logical -- default \code{FALSE}; print function-associated messages to the console.}
}
\value{
A \code{flowstate} containing spectrally-associated bead/cellular events.
}
\description{
This function is entirely dependent -- as of now -- on the following:
\itemize{
\item Spectroflo software raw reference controls  -- tested using SpectroFlo 3.3.0.
\item Spectroflo software naming convention:
\itemize{
\item Directory -- './Raw/Reference Group/...'
\item File names -- 'MARKER FLUOR (Beads|Cells).fcs' (literal space characters separating)
\item Quality reference controls -- well resolved, dominant scatter populations
}
}
}
\examples{
\dontrun{
##single-color reference controls
raw.ref.files <- system.file("extdata", package = "flowstate") |>
list.files(full.names = TRUE, pattern = "Beads")

##simulate a SpectroFlo directory structure
raw.ref.dir <- file.path(tempdir(),"Raw","Reference Group")
if(!dir.exists(raw.ref.dir)) dir.create(raw.ref.dir,recursive = TRUE)
file.copy(
  from = raw.ref.files,
  to = file.path(raw.ref.dir,basename(raw.ref.files))
) |> invisible()

##generate spectral events
ref <- reference.group.spectral.events(
  raw.reference.group.directory = raw.ref.dir,
  cluster.types = 'beads',
  plot.select = TRUE,
  verbose = TRUE
)

##a flowstate containing spectral events
class(ref)
ref[['data']]

##peak detectors
ref$data[,.SD,.SDcols = is.factor] |> unique()

##number of spectral events as defined using defaults
ref$data[,.N,by = sample.id]

##saved plot output
list.files(tempdir()) |> grep(pattern = "select.*.pdf",value = TRUE)

##detector.override
##for the sake of example
ref <- reference.group.spectral.events(
  raw.reference.group.directory = raw.ref.dir,
  cluster.types = 'beads',
  detector.override = c("Unstained (Beads)" = "UV7")
)
ref$data[,.SD,.SDcols = is.factor] |> unique()

##n.spectral.events.override
##for the sake of example
ref <- reference.group.spectral.events(
  raw.reference.group.directory = raw.ref.dir,
  cluster.types = 'beads',
  n.spectral.events.override = c("TNFa PE (Beads)" = 500)
)

##number of spectral events as defined using n.spectral.events.override
ref$data[,.N,by = sample.id]
}

}
