% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_flowstate.R
\name{write.flowstate}
\alias{write.flowstate}
\title{Write a \code{flowstate} as a Flow Cytometry Standard file (FCS 3.1)}
\usage{
write.flowstate(
  flowstate.object,
  new.fil = NULL,
  add.fil.mod = TRUE,
  file.dir,
  endianness = c("little", "big")
)
}
\arguments{
\item{flowstate.object}{A \code{flowstate}}

\item{new.fil}{Character string -- default \code{NULL}; if defined, the keyword \code{'$FIL'} will be updated with the provided value.}

\item{add.fil.mod}{Logical -- default \code{TRUE}; the suffix \code{'flowstateMOD'}(string) will be appended to the value of \code{'$FIL'}.}

\item{file.dir}{Character vector -- a \link[=file.path]{base}; a file directory for saving the output FCS 3.1 file.}

\item{endianness}{Character string -- default \code{"little"}; see \link[base:readBin]{endian}.}
}
\value{
The FCS 3.1 file is written to disk and a summary message is displayed.
}
\description{
A \code{flowstate} -- following \href{https://pmc.ncbi.nlm.nih.gov/articles/PMC2892967/}{Flow Cytometry Standard} conventions -- is converted to FCS 3.1 and written to disk. The conversion is as follows:
\itemize{
\item \verb{[['keywords']]}, \verb{[['parameters']]}, and \verb{[['spill']]} (if present) are converted to character string using a delimiter \code{"|"} to separate keyword-value pairs.
\itemize{
\item this string forms the \code{TEXT} segment of the FCS file.
}
\item \verb{[['data']]} is written as binary.
\itemize{
\item this forms the \code{DATA} segment of the FCS file.
}
\item The \code{HEADER} segment (string) -- containing the required offsets -- is derived from \code{TEXT} and \code{DATA}.
}

The FCS 3.1 compliant file is then written to disk.

\emph{N.B.: During testing, FCS 3.1 files written to disk by \link{write.flowstate} were fully compatible with \link{read.flowstate} and \link[flowCore:read.FCS]{flowCore::read.FCS}; issues were only with proprietary software (i.e., FlowJo). If proprietary software is to be used with FCS 3.1 files generated by \link{write.flowstate}, caution is advised.}
}
